Establish token provider trait
