+++ aws/rust-runtime/aws-credential-types/src/lib.rs
@@ -23,5 +23,7 @@
 pub mod credential_fn;
 mod credentials_impl;
 pub mod provider;
+mod token_impl;
 
 pub use credentials_impl::Credentials;
+pub use token_impl::AccessToken;
